% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectivity_score.R
\name{CRN}
\alias{CRN}
\title{Correlation based network}
\usage{
CRN(x)
}
\arguments{
\item{x}{microarray dataset with genes in columns and samples in rows.}
}
\value{
s matrix containing connectivity scores
}
\description{
Correlation based network
}
\examples{
library(EGRNi)
data(gene_exp)
t_geneexp<-t(gene_exp)
CRN(t_geneexp)
}
