% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Edgescore.R
\name{Edgescore}
\alias{Edgescore}
\title{Edge score obtained from 4 different methods for Ensemble Gene Regulatory Network Inference}
\usage{
Edgescore(path)
}
\arguments{
\item{path}{path to file name}
}
\value{
a \code{tibble}
}
\description{
Edge score obtained from 4 different methods for Ensemble Gene Regulatory Network Inference
}
\references{
Sarkar, C., Parsad, R., Mishra, D.C. and Rai, A (2020). An ensemble approach for gene regulatory network study in rice blast. Journal of Crop and Weed , 16 , 1-8.
}
\author{
Chiranjib Sarkar(cschiranjib9@gmail.com)
}
