% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectivity_score.R
\name{PCN}
\alias{PCN}
\title{Principal component regression based network}
\usage{
PCN(x)
}
\arguments{
\item{x}{microarray dataset with genes in columns and samples in rows.}
}
\value{
s matrix containing connectivity scores
}
\description{
Principal component regression based network
}
\examples{
library(EGRNi)
data(gene_exp)
t_geneexp<-t(gene_exp)
PCN(t_geneexp)
}
\references{
Sarkar, C., Parsad, R., Mishra, D.C. and Rai, A (2020). An ensemble approach for gene regulatory network study in rice blast. Journal of Crop and Weed , 16 , 1-8.
}
\author{
Chiranjib Sarkar(cschiranjib9@gmail.com)
}
