% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectivity_score.R
\name{PLSN}
\alias{PLSN}
\title{Partial least square based network}
\usage{
PLSN(x)
}
\arguments{
\item{x}{microarray dataset with genes in columns and samples in rows.}
}
\value{
s matrix containing connectivity scores
}
\description{
Partial least square based network
}
\examples{
library(EGRNi)
data(gene_exp)
t_geneexp<-t(gene_exp)
PLSN(t_geneexp)
}
