% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectivity_score.R
\name{RidgN}
\alias{RidgN}
\title{Ridge regression based network}
\usage{
RidgN(x)
}
\arguments{
\item{x}{microarray dataset with genes in columns and samples in rows.}
}
\value{
s matrix containing connectivity scores
}
\description{
Ridge regression based network
}
\examples{
library(EGRNi)
data(gene_exp)
t_geneexp<-t(gene_exp)
RidgN(t_geneexp)
}
